Refactor app_bsky::notification
