--- atprotolib-rs/src/types/app_bsky/notification/get_unread_count.rs
@@ -1,13 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/*
-    app.bsky.notification.getUnreadCount
-*/
-
-/// The response to a request for the unread count of notifications.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GetUnreadCountResponse {
-    /// The count of unread notifications.
-    #[serde(rename = "count", default)]
-    pub count: i32
-}
