+++ atprotolib-rs/src/types/app_bsky/notification/mod.rs
@@ -1,27 +1,10 @@
+pub mod api_requests;
+pub mod api_responses;
+
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
 use crate::types::{app_bsky::actor::ProfileView, com_atproto::label::Label};
-
-/*
-    app.bsky.notification.listNotifications
-*/
-
-/// The response to a request for a user's notifications.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct ListNotificationsResponse {
-    /// The cursor for the stream.
-    cursor: Option<String>,
-
-    /// A list of the user's notifications.
-    notifications: Vec<Notification>,
-
-    /// Whether the notifications are priority notifications.
-    priority: bool,
-
-    /// The date and time the notifications were last seen.
-    seen_at: Option<DateTime<Utc>>
-}
 
 /// Represents a notification.
 #[derive(Serialize, Deserialize, Debug)]
@@ -45,7 +28,7 @@ pub struct Notification {
     /// The subject of the reason.
     #[serde(rename = "reasonSubject", skip_serializing_if = "Option::is_none")]
     pub reason_subject: Option<String>,
-    
+
     /// The record associated with the notification.
     #[serde(rename = "record")]
     pub record: serde_json::Value,
